Move AST into separate file
